Fix Javadoc generation by moving validation logic out of sql/api
The documentation build was failing with "illegal combination of modifiers:
abstract and static" errors when generating Java stubs. This was caused by
a private[sql] companion object in the sql/api DataStreamReader interface.

Changes:
- Remove companion object from sql/api DataStreamReader (API interfaces
  should not contain implementation logic)
- Add StreamingSourceValidation object to sql/catalyst for shared validation
  logic used by Classic implementation
- Add local validation to sql/connect DataStreamReader companion object
  (Connect doesn't depend on catalyst, so needs its own copy)
- Update imports in Classic and Connect implementations with proper
  alphabetical ordering

The validation logic is now properly separated from the API definition,
allowing Javadoc to generate correctly without creating invalid Java code.
