+++ sql/core/src/main/scala/org/apache/spark/sql/classic/DataStreamReader.scala
@@ -22,13 +22,13 @@ import scala.jdk.CollectionConverters._
 import org.apache.spark.annotation.{Evolving, Experimental}
 import org.apache.spark.sql.catalyst.analysis.{NamedStreamingRelation, UnresolvedRelation}
 import org.apache.spark.sql.catalyst.plans.logical.UnresolvedDataSource
+import org.apache.spark.sql.catalyst.streaming.StreamingSourceValidation.validateSourceName
 import org.apache.spark.sql.catalyst.util.{CaseInsensitiveMap, CharVarcharUtils}
 import org.apache.spark.sql.classic.ClassicConversions._
 import org.apache.spark.sql.errors.QueryCompilationErrors
 import org.apache.spark.sql.execution.datasources.json.JsonUtils.checkJsonSchema
 import org.apache.spark.sql.execution.datasources.xml.XmlUtils.checkXmlSchema
 import org.apache.spark.sql.streaming
-import org.apache.spark.sql.streaming.DataStreamReader.validateSourceName
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
